requirements.txt
Django==2.0.3

django-crispy-forms==1.7.2

django-filter==1.1.0

Pillow==5.0.0

pytz==2018.3
